cjdns-bencode: Replace serde/bendy with native bencode library.
Cjdns uses bencode for inter-router communication, but because
it builds messages backwards (data-first, headers pushed one by one)
it requires the content to have a round length in order to maintain
4 byte alignment of the header structs.

The original solution to this was a hack, based on the fact that
cjdns bencoded messages are always of type Dict. The length of the
first key in the dict is zero-padded to make the entire message
length a multiple of four.

For example:

d7:address78:v21.0000.0aa2.7ac1.8a8b......e

becomes:

d0007:address78:v21.0000.0aa2.7ac1.8a8b......e

This however violates the bencode spec and fails validation with
a standard bencode parser, so until now, we used a patched version
of bendy to do the bencode parsing.

Maintaining a fork of bendy is far from ideal, and the bendy/serde
combination was unable to process binary in bencoded strings,
severely limiting the usefulness of the cjdns-admin crate.

This patch eliminates bendy and migrates to a port of the cjdns
bencode parser to Rust. The serializer builds the message backwards
making it compatible with cjdns C code, so the same parser can be
used for both. Also a bencode object implementation is added which
has convenient list/dict accessors and conversion functions, and
all uses of serde are replaced with cjdns_bencode::object::*